\name{get_scalar_samples}
\alias{get_scalar_samples}
\title{
Extract Samples for Specified Scalar Model Parameter 
}
\description{
This function takes an output of \code{clustGLMM} and finds all values related to specified model parameter in a given range (possibly thinned). The implementation reacts not only to what is given but also to what is NOT given. To use it appropriately you need to know which model parameters are group-specific, outcome-dependent and many dimensions do they have.
}
\usage{
get_scalar_samples(mcmc, what, gspec, dimspec, yspec, 
                   burnin = 0, thin = 1, 
                   iterations, chains)
}
\arguments{
  \item{mcmc}{
an output of \code{\link{clustGLMM}} accepted in both formats (list or matrix).
}
  \item{what}{
a name of the parameter to be extracted; see \code{names(mcmc$save)} for possible options. 
}
  \item{gspec}{
a group specifier (values in \code{1:mcmc$G}); if not group-specific parameter this \emph{should be missing}.
}
  \item{dimspec}{
a dimension specifier; not specifying this is possible only when the parameter is scalar.
}
  \item{yspec}{
an outcome specifier; if the parameter is not outcome-dependent then this \emph{should be missing}.
}
  \item{burnin}{
the length of burn-in period (default 0), i.e. number of initial samples to ignore for inference.
}
  \item{thin}{
thinning rate (default 1); only every \code{thin}th value is used for inference (must be a positive integer). 
}
  \item{iterations}{
a vector of iterations numbers to extract, overrides any given \code{burnin} or \code{thin}. 
}
  \item{chains}{
a vector of chain numbers to be used.
}
}
\details{
Function \code{get_scalar_samples} is used mainly internally in plotting functions to find appropriate values given the name of the parameter \code{what}, group specification \code{gspec} (a number of a cluster), outcome specification \code{yspec} (name of the outcome) and dimension specification \code{dimspec}. For vector parameters \code{dimspec} is just a single number, if \code{what} is a matrix parameter then 2 numbers are expected.

\code{iterations} declares which \code{m} values should be gathered. 
If \code{iterations} is missing, the returned values are regulated by changing \code{burnin} and \code{thin}.
First, only iterations with \code{m > burnin} are selected, then, from the remaining values only \code{seq(1, total_remaining_count, by=thin)} are selected. Also, \code{\link{post_processing}} could turn some of the cluster-specific variables into NA, hence, if \code{gspec} is missing, these cases are avoided.
}
\value{
A list containing \code{whatLAB} (the default label for this scalar parameter) and \code{draws} a list with the draws.

\code{draws} is a list (for each element of \code{chains}) containing \code{data.frame} of three columns: 
\describe{
  \item{m}{iteration numbers,}
  \item{chain}{the chain number,}
  \item{y}{the values of the required parameter.}
}
}
\references{
Vávra, J., Komárek, A., Grün, B., Malsiner-Walli, G. Clusterwise multivariate regression of mixed-type panel data. Stat Comput 34, 46 (2024). https://doi.org/10.1007/s11222-023-10304-5
}
\author{
Jan Vávra
}
\seealso{
\code{\link{clustGLMM}}, \code{\link{plot.clustglmm}}
}
\examples{

### Get some longitudinal mixed type data
### here generated by generate_longitudinal_mixed_type_data()
data("longitudinal_mixed_type_data")

# Examples here demonstrate the basic use of the available functions on a toy dataset.
# For more thorough analyses see the vignettes.


###-----------------------------------------------------------------------------
###     Prepare inputs for clustGLMM()
###-----------------------------------------------------------------------------
id <- "i"

family <- c("num", "poi", "bin", "ord", "cat")
names(family) <- Ys <- c("ynum", "ypoi", "ybin", "yord", "ycat")

## Create list of formulae for each outcome
formula <- list()
for(y in Ys){
  formula[[y]] <- list()
  formula[[y]]$fixed <- ~ f
  formula[[y]]$group <- ~ x
  formula[[y]]$random <- ~ 1
  formula[[y]]$offset <- ""
}
formula
# intercept term will be group-specific 
# since it appears in $group which has a preference over $fixed

## All other inputs need to be prepared,
## unless the default values are desired (varying, save, howsave, param, tuning).
## See the vignette for how these can be adjusted.

###-----------------------------------------------------------------------------
###     Calling clustGLMM()
###-----------------------------------------------------------------------------

Gmax <- 5     # maximal number of mixture components to be considered
nchains <- 2  # number of chains to be sampled 

## No inits specified
set.seed(123456789)
mcmc <- clustGLMM(formula = formula, id = id, family = family, 
                  data = longitudinal_mixed_type_data, 
                  G = Gmax, iter = 100, nchains = nchains)

###-----------------------------------------------------------------------------
###     The use of get_scalar_parameter_samples
###-----------------------------------------------------------------------------
# Finding values for a specific parameter
# beta_num is group-specific, outcome-dependent and is a vector
get_scalar_samples(mcmc, what = "beta_num", gspec = 2, dimspec = 1, yspec = "ynum", 
                   burnin = 0, thin = 1, chains = 1:nchains)

# beta_num_fix is not group-specific, but outcome-dependent and is a vector
get_scalar_samples(mcmc, what = "beta_num_fix", dimspec = 1, yspec = "ynum",
                   burnin = 0, thin = 1, chains = 1:nchains)

mcmc$varying["InvSigma"]
# InvSigma is not group-specific, not outcome-dependent, but is a symmetric matrix    
# only elements [i,j] where i<=j are allowed
get_scalar_samples(mcmc, what = "InvSigma", dimspec = c(1,1), 
                   burnin = 50, thin = 1, chains = 1:nchains)

# corSigma is not group-specific, not outcome-dependent, but is a symmetric matrix without diagonal
# only elements [i,j] where i<j are allowed
get_scalar_samples(mcmc, what = "corSigma", dimspec = c(1,2), 
                   burnin = 50, thin = 10, chains = 1:nchains)

}

