\name{get_Sigma}
\alias{get_Sigma}
\alias{get_sd_num}
\title{
Extract the m-th Sample of Sigma or sd_num
}
\description{
These two parameters are saved under different formats or tranformation.
These functions return the m-th sample of the corresponding parameter even if it has to be transformed from another equivalent saved parameter.
}
\usage{
get_Sigma(draws, m, settings, g = NA, howsave = "data.frame", 
          Sigmacols, InvSigmacols, sdSigmacols, corSigmacols)  
get_sd_num(draws, m, settings, y = NA, howsave = "data.frame") 
}
\arguments{
  \item{draws}{
samples of model parameters in either list or data.frame (matrix) structure.
}
  \item{m}{
an iteration to choose.
}
  \item{settings}{
a matrix of model settings (in output of \code{clustglmm}).
}
  \item{g}{
a cluster label.
}
  \item{y}{
a name of an outcome.
}
  \item{howsave}{
the format under which \code{draws} is saved.
}
  \item{Sigmacols}{
columns in \code{draws} where Sigma matrix is stored, found automatically if missing.
}
  \item{InvSigmacols}{
columns in \code{draws} where InvSigma matrix is stored, found automatically if missing.
}
  \item{sdSigmacols}{
columns in \code{draws} where sdSigma is stored, found automatically if missing.
}
  \item{corSigmacols}{
columns in \code{draws} where corSigma matrix is stored, found automatically if missing.
}
}
\details{
From \code{draws}, only the \code{m}-th element is used, which slightly differs depending on \code{howsave}.
Functions are able to detect group-specificity and dimension from \code{settings}.

If \code{Sigma} is group-specific, then only matrix for gorup \code{g} is returned.
If \code{Sigma} is not saved among draws, then it is reconstructed from \code{InvSigma} and inversion.
If neither of \code{Sigma} and \code{InvSigma} is saved among draws, then it is reconstructed from standard deviations \code{sdSigma} or \code{corSigma}.
If none of the above, a warning is triggered and diagonal unit matrix is returned. 

The columns where the results are stored are found through function \code{\link[base]{grep}} with pattern composed of the name of the parameter and the cluster label if given.
This could be time-consuming, especially, when repeating over many \code{m} values.
Hence, there is an opportunity to supply the columns via \code{Sigmacols}, \code{InvSigmacols}, \code{sdSigmacols} and \code{corSigmacols}.

On the other hand \code{get_sd_num} returns standard deviations for all clusters if cluster-specific. 
First, it tries to get it from \code{sd_num}, if saved.
If not it transforms either from \code{prec_num} or \code{var_num}. 
If none of the above, error is triggered. 
}
\value{
A single covariance matrix \code{Sigma} or standard deviation(s) of numeric outcome \code{y}.
}
\references{
Vávra, J., Komárek, A., Grün, B., Malsiner-Walli, G. Clusterwise multivariate regression of mixed-type panel data. Stat Comput 34, 46 (2024). https://doi.org/10.1007/s11222-023-10304-5
}
\author{
Jan Vávra
}
\seealso{
\code{\link{clustGLMM}}
}
\examples{

### Get some longitudinal mixed type data
### here generated by generate_longitudinal_mixed_type_data()
data("longitudinal_mixed_type_data")

# Examples here demonstrate the basic use of the available functions on a toy dataset.
# For more thorough analyses see the vignettes.


###-----------------------------------------------------------------------------
###     Prepare inputs for clustGLMM()
###-----------------------------------------------------------------------------
id <- "i"

family <- c("num", "poi", "bin", "ord", "cat")
names(family) <- Ys <- c("ynum", "ypoi", "ybin", "yord", "ycat")

## Create list of formulae for each outcome
formula <- list()
for(y in Ys){
  formula[[y]] <- list()
  formula[[y]]$fixed <- ~ f
  formula[[y]]$group <- ~ x
  formula[[y]]$random <- ~ 1
  formula[[y]]$offset <- ""
}
formula
# intercept term will be group-specific 
# since it appears in $group which has a preference over $fixed

## All other inputs need to be prepared,
## unless the default values are desired (varying, save, howsave, param, tuning).
## See the vignette for how these can be adjusted.

###-----------------------------------------------------------------------------
###     Calling clustGLMM()
###-----------------------------------------------------------------------------

Gmax <- 5     # maximal number of mixture components to be considered
nchains <- 2  # number of chains to be sampled 

## No inits specified
set.seed(123456789)
mcmc <- clustGLMM(formula = formula, id = id, family = family, 
                  data = longitudinal_mixed_type_data, 
                  G = Gmax, iter = 100, nchains = nchains)
                       
###-----------------------------------------------------------------------------
###     Getting Sigma matrix and sd_num
###-----------------------------------------------------------------------------

# We take the last generated value by m = mcmc$iter

Sigma <- get_Sigma(mcmc$draws[[1]], mcmc$iter, mcmc$settings[[1]],
                   1, howsave = mcmc$howsave)
sd_num <- get_sd_num(mcmc$draws[[1]], mcmc$iter, mcmc$settings[[1]],
                     "ynum", howsave = mcmc$howsave)                  

}
\keyword{internal}
