\name{plot.clustglmm}
\alias{plot.clustglmm}
\title{
Plot the Output of \code{clustGLMM}
}
\description{
Triggers plotting functions for selected parameter depending on the choice of the plot.
}
\usage{
\method{plot}{clustglmm}(x, what="ng", which="traceplots", ng_trace_chain_split=TRUE, ...)
}
\arguments{
  \item{x}{
an object of class \code{"clustglmm"}, an output of function \code{clustGLMM}.
}
  \item{what}{
a string containing the name of the parameter. 
}
  \item{which}{
a string declaring the type of the plot to be created.
}
  \item{ng_trace_chain_split}{a logical indicating ...}
  \item{...}{
other arguments.
}
}
\details{
Depending on \code{which} the following functions are called:
\describe{
  \item{ACF}{creates auto-correlation function plot, calls \code{\link{plot_ACF}} or \code{\link{plot_ACF_param}}.}
  \item{ECDF}{creates empirical cumulative distribution function plot, calls \code{\link{plot_ECDF}} or \code{\link{plot_ECDF_param}}.}
  \item{kerneldensity}{creates kernel density estimator plot, calls \code{\link{plot_kerneldensity}} or \code{\link{plot_kerneldensity_param}}.}
  \item{traceplots}{creates traceplots (samples vs iteration), calls \code{\link{plot_traceplots}} or \code{\link{plot_traceplots_param}}.}
  \item{clusters_ECDF}{creates empirical cumulative distribution function plot where different clusters are compared (\code{what} needs to be a cluster-specific parameter), calls \code{\link{plot_clusters}} or \code{\link{plot_clusters_ECDF_param}}.}
  \item{clusters_kerneldensity}{creates kernel density estimator plot where different clusters are compared (\code{what} needs to be a cluster-specific parameter), calls \code{\link{plot_clusters}} or \code{\link{plot_clusters_kerneldensity_param}}.}
}

If the user tries to specify one of: \code{gspec} or \code{dimspec}, it is treated as an attempt to plot only a single scalar parameter (functions without \code{..._param}).
Be careful to specify the dimensions appropriately (including \code{yspec}). 
Inner function \code{\link{get_scalar_samples}} triggers a warning or even an error if misspecified.

Otherwise if neither \code{gspec} nor \code{dimspec} are given, then \code{..._param} function is triggered.
It is designed to create plots for all chains and all dimensions related to the given \code{what} model parameter.
The output is organized into multiple carefully arranged layout.
The only way to plot a smaller subset of parameters is via \code{yspec} argument, which accepts even multiple outcome values.
However, only plots for an intersection of given and plausible outcomes for the given parameter are returned.


These functions except also \code{burnin} (initial part of the chain to be ignored), \code{thin} (plot only every \code{thin}th sample) and \code{chains} (a vector of chain numbers to be plotted).
}
\value{
  No return value, only creates plots.
}
\references{
Vávra, J., Komárek, A., Grün, B., Malsiner-Walli, G. Clusterwise multivariate regression of mixed-type panel data. Stat Comput 34, 46 (2024). https://doi.org/10.1007/s11222-023-10304-5
}
\author{
Jan Vávra
}
\seealso{
\code{\link{clustGLMM}}
}
\examples{

### Get some longitudinal mixed type data
### here generated by generate_longitudinal_mixed_type_data()
data("longitudinal_mixed_type_data")

# Examples here demonstrate the basic use of the available functions on a toy dataset.
# For more thorough analyses see the vignettes.


###-----------------------------------------------------------------------------
###     Prepare inputs for clustGLMM()
###-----------------------------------------------------------------------------
id <- "i"

family <- c("num", "poi", "bin", "ord", "cat")
names(family) <- Ys <- c("ynum", "ypoi", "ybin", "yord", "ycat")

## Create list of formulae for each outcome
formula <- list()
for(y in Ys){
  formula[[y]] <- list()
  formula[[y]]$fixed <- ~ f
  formula[[y]]$group <- ~ x
  formula[[y]]$random <- ~ 1
  formula[[y]]$offset <- ""
}
formula
# intercept term will be group-specific 
# since it appears in $group which has a preference over $fixed

## All other inputs need to be prepared,
## unless the default values are desired (varying, save, howsave, param, tuning).
## See the vignette for how these can be adjusted.

###-----------------------------------------------------------------------------
###     Calling clustGLMM()
###-----------------------------------------------------------------------------

Gmax <- 5     # maximal number of mixture components to be considered
nchains <- 2  # number of chains to be sampled 

## No inits specified
set.seed(123456789)
mcmc <- clustGLMM(formula = formula, id = id, family = family, 
                  data = longitudinal_mixed_type_data, 
                  G = Gmax, iter = 100, nchains = nchains)
                  
###-----------------------------------------------------------------------------
###     Using plot() function on clustglmm object
###-----------------------------------------------------------------------------

# traceplots by default                       
plot(mcmc)
plot(mcmc, "Gplus")
plot(mcmc, "beta_num", burnin=10, thin=2, chains=c(1))

# change the plot type
plot(mcmc, "w", "ACF")
plot(mcmc, "sdSigma", "ECDF")
plot(mcmc, "corSigma", "kerneldensity")
plot(mcmc, "beta_bin", "clusters_kerneldensity")
plot(mcmc, "beta_ord", "clusters_ECDF")

}
