\name{plot_cat_vs_x_grouped}
\alias{plot_cat_vs_x_grouped}
\alias{plot_num_vs_x_grouped}
\title{
Plot Clustered Longitudinal (Panel) Data 
}
\description{
Functions that work with longitudinal data and plot the outcomes vs a covariate with respect to some grouping, e.g. clustering found by \code{\link{clustGLMM}}. 
}
\usage{
plot_num_vs_x_grouped(data, y = "y", x = "x", group, id = "i", units, 
                      xbreaks, add_lowess = TRUE, alpha = 0.7,
                      legend_placement = "topright")
plot_cat_vs_x_grouped(data, y = "y", x = "x", group, 
                      xbreaks, main = NA, margin = 0.05, reverse = FALSE)
}
\arguments{
  \item{data}{
a \code{data.frame} containing \code{y}, \code{x}, \code{group}, \code{id}.
}
  \item{y}{
a name of the outcome to be plotted.
}
  \item{x}{
a name of numeric or factor regressor. 
}
  \item{group}{
a name of the clustering variable that divides different ids into groups.
If missing, all observations are given the label "0" plotted with \code{\link{slategray.colors}} and \code{legend_placement} is set by default to \code{NULL}.
}
  \item{id}{
a name of the id variable.
}
  \item{units}{
a vector of unit ids to be plotted, if missing all ids are plotted.
}
  \item{xbreaks}{
a set of breaks along which to cut regressor \code{x}, if missing \code{x} is divided equidistantly into 10 categories.
}
  \item{add_lowess}{
a Boolean whether to add lowess curves for each group label.
}
  \item{alpha}{
a color transparency factor; applied only for overlapping boxplots if covariate \code{x} is \code{factor}. 
}
  \item{main}{
plot title to be printed above the plot.
}
  \item{margin}{
how far from each other should different groups be displayed.
}
  \item{reverse}{
a Boolean whether to reverse the order of ordinal categories; if \code{TRUE} 0 on top, if \code{FALSE} 0 at the bottom.
}
  \item{legend_placement}{
a string describing the position of legend (see \code{\link[graphics]{legend}}) or \code{NULL} to omit the legend.
}
}
\details{
Colors are created automatically by \code{rainbow_hcl} from package \code{colorspace}.
\code{\link{slategray.colors}} are used for group 0 (unclassified).

Output of \code{plot_num_vs_x_grouped} depends on the type of covariate \code{x}:
\describe{
\item{factor}{boxplots of the outcome \code{y} wrt levels of covariate \code{x} with a transparent color
determined by \code{group}. Boxplots for the same group may overlap, hence, transparent colors.}
\item{numeric}{spaghetti-plots the outcome \code{y} wrt covariate \code{x} coloured by \code{group}. 
Only units included in \code{units} are plotted. }
}

If \code{x} is not already a factor, \code{plot_cat_vs_x_grouped} divides the data by a numeric covariate \code{x} into groups divided by \code{xbreaks}. 
Then, for each group label in \code{group} (including group 0 = unclassified) a spineplot capturing the evolution of proportions of a categorical outcome \code{y} wrt the covariate \code{x}. 
}
\value{
The functions do not return anything but produce a plot.
}
\references{
Vávra, J., Komárek, A., Grün, B., Malsiner-Walli, G. Clusterwise multivariate regression of mixed-type panel data. Stat Comput 34, 46 (2024). https://doi.org/10.1007/s11222-023-10304-5
}
\author{
Jan Vávra
}
\seealso{
\code{\link{clustGLMM}}
}
\examples{

### Get some longitudinal mixed type data
### here generated by generate_longitudinal_mixed_type_data()
data("longitudinal_mixed_type_data")

# Examples here demonstrate the basic use of the available functions on a toy dataset.
# For more thorough analyses see the vignettes.


###-----------------------------------------------------------------------------
###     Prepare inputs for clustGLMM()
###-----------------------------------------------------------------------------
id <- "i"

family <- c("num", "poi", "bin", "ord", "cat")
names(family) <- Ys <- c("ynum", "ypoi", "ybin", "yord", "ycat")

## Create list of formulae for each outcome
formula <- list()
for(y in Ys){
  formula[[y]] <- list()
  formula[[y]]$fixed <- ~ f
  formula[[y]]$group <- ~ x
  formula[[y]]$random <- ~ 1
  formula[[y]]$offset <- ""
}
formula
# intercept term will be group-specific 
# since it appears in $group which has a preference over $fixed

## All other inputs need to be prepared,
## unless the default values are desired (varying, save, howsave, param, tuning).
## See the vignette for how these can be adjusted.

###-----------------------------------------------------------------------------
###     Calling clustGLMM()
###-----------------------------------------------------------------------------

Gmax <- 5     # maximal number of mixture components to be considered
nchains <- 2  # number of chains to be sampled 

## No inits specified
set.seed(123456789)
mcmc <- clustGLMM(formula = formula, id = id, family = family, 
                  data = longitudinal_mixed_type_data, 
                  G = Gmax, iter = 100, nchains = nchains)
                       
###-----------------------------------------------------------------------------
###     Clustering based on not post-processed data (using sampled indicators)
###-----------------------------------------------------------------------------
### Take the sampled U indicators
threshold <- 0.5 # if not exceeded, unit remains unclassified

# posterior mode of Gplus=number of nonemtpy components
mcmc$modeGplus

# labels of non-empty clusters in each chain
mcmc$clusters

# clustering based on sampled indicators
myclustering <- mcmc$clustering

# the highest estimated clustering probability among all clusters
mcmc$certainty

# units with frequency ratio smaller than chosen threshold remain unclassified (group 0)
myclustering[mcmc$certainty < threshold] <- 0
clusters <- list()

# proposed clustering by each chain separately
for(ch in 1:mcmc$nchains){
  longitudinal_mixed_type_data[,paste0("clustering",ch)] <- 
    myclustering[longitudinal_mixed_type_data[,id],ch]
  longitudinal_mixed_type_data[,paste0("certainty",ch)] <- 
    mcmc$certainty[longitudinal_mixed_type_data[,id],ch]
}

data1 <- longitudinal_mixed_type_data[longitudinal_mixed_type_data$j==1,]
# Confusion matrices for each chain separately:
table(data1$clustering1, data1$g)
# We can also see the suitable permutation for labels 

###-----------------------------------------------------------------------------
###     Plotting the resulting clustering
###-----------------------------------------------------------------------------

# Plotting numeric outcomes
plot_num_vs_x_grouped(longitudinal_mixed_type_data, y = "ynum", x = "x", 
                      group = "g", id = "i", units = 1:mcmc$n)
                      
plot_num_vs_x_grouped(longitudinal_mixed_type_data, y = "ynum", x = "x",
                      group = "clustering1", id = "i", units = 1:mcmc$n)

# Plotting binary, ordinal, categorical outcomes
plot_cat_vs_x_grouped(longitudinal_mixed_type_data, y = "ybin", x = "x",
                      group = "g")
                      
plot_cat_vs_x_grouped(longitudinal_mixed_type_data, y = "ybin", x = "x",
                      group = "g", reverse = TRUE)
                      
plot_cat_vs_x_grouped(longitudinal_mixed_type_data, y = "ybin", x = "x",
                      group = "clustering1")
                      
plot_cat_vs_x_grouped(longitudinal_mixed_type_data, y = "yord", x = "x",
                      group = "g", reverse = TRUE)
                      
plot_cat_vs_x_grouped(longitudinal_mixed_type_data, y = "yord", x = "x",
                      group = "g", reverse = FALSE)
                      
plot_cat_vs_x_grouped(longitudinal_mixed_type_data, y = "yord", x = "x",
                      group = "clustering1")
                      
plot_cat_vs_x_grouped(longitudinal_mixed_type_data, y = "ycat", x = "x",
                      group = "g")
                      
plot_cat_vs_x_grouped(longitudinal_mixed_type_data, y = "ycat", x = "x",
                      group = "clustering1")
}
