\name{psurvey}
\alias{psurvey}
\alias{psurvey_latent}
\docType{data}
\title{
A Simulated Panel Survey Data
}
\description{
An artificially created dataset to demonstrate the use of \code{\link{clustGLMM}}. It contains data about 500 people (\code{i}) each observed random number of times in a span of 20 years. The data were generated by assuming 3 latent groups (\code{g}). People in group 1 have lower income (did not finish university) than other two groups. Groups 2 and 3 are similar in income but differ in the attitude, while people in group 2 have more relaxing approach and spend money on trips, people in group 3 work hard on their career and enjoy life less.
%1 (people who did not reach any university degree - lower income, more trips earlier, general satisfation), 
%2 (people who reached university degree with relaxing aproach to life - higher income, more trips after studies, higher satisfaction ) 
%3 (hard-working people who have a successful career - slightly higher income than 2, less trips, less satisfaction).
}
\usage{
data("psurvey")
data("psurvey_latent")
}
\format{
  \code{psurvey} is a data frame with 4512 observations on the following 6 variables
  \describe{
    \item{\code{i}}{a numeric vector - id of a person}
    \item{\code{j}}{a numeric vector - order of observation for a person}
    \item{\code{age}}{a numeric vector - age of a person when observation taken}
    \item{\code{income}}{a numeric vector - income of a person in artificial units}
    \item{\code{ntrips}}{a numeric vector - number of trips taken in the last year}
    \item{\code{satisfaction}}{an ordered factor with levels \code{0} < \code{1} < \code{2} - level of self-evaluated satisfaction with their current lifestyle (0 not satisfied, 1 neither-nor, 2 satisfied)}
  }
  
  Additionally, data frame \code{psurvey_latent} contains columns that are known from the creation process, but should not be used for analysis:
  \describe{
    \item{\code{i}}{a numeric vector - id of a person}
    \item{\code{j}}{a numeric vector - order of observation for a person}
    \item{\code{g}}{a numeric vector - a true group label from 1, 2, 3}
    \item{\code{b_num}}{a numeric vector - true random intercept for \code{income}}
    \item{\code{b_poi}}{a numeric vector - true random intercept for \code{ntrips}}
    \item{\code{b_ord}}{a numeric vector - true random intercept for \code{satisfaction}}
  }
}
\examples{
data(psurvey)
head(psurvey)
summary(psurvey)

data(psurvey_latent)
head(psurvey_latent)
}
\keyword{datasets}
