\name{rescale_matrix}
\alias{rescale_matrix}
\alias{rescale_list}
\title{
Rescale the Raw Draws 
}
\description{
In \code{\link{clustGLMM}} if \code{standardize=TRUE}, the model parameters are sampled using the scaled numeric outcomes and regressors. The goal of these functions is to transform these parameters into the original scale. Works for a specific chain only.
}
\usage{
rescale_matrix(draws, settings, family, lfixnames, lgrpnames, lrannames, 
               Kord, Kcat, centers, scales)
rescale_list(draws, settings, family, lfixnames, lgrpnames, lrannames, 
             Kord, Kcat, centers, scales)
}
\arguments{
  \item{draws}{
an element of \code{clustGLMM$draws} for a certain chain, either a list or a matrix.
}
  \item{settings}{
a matrix containing the information about model parameters.
}
  \item{family}{
a vector of distributional families named by the outcome name in \code{data}.
}
  \item{lfixnames}{
a list of \code{fixed} regressors for each outcome. 
}
  \item{lgrpnames}{
a list of \code{group} regressors for each outcome. 
}
  \item{lrannames}{
a list of \code{random} regressors for each outcome. 
}
  \item{Kord}{
a named vector of number of levels (minus 1) for ordinal outcomes.
}
  \item{Kcat}{
a named vector of number of levels (minus 1) for categorical outcomes.
}
  \item{centers}{
a named vector of centers of numeric outcomes and regressors.
}
  \item{scales}{
a named vector of scales of numeric outcomes and regressors.
}
}
\details{
Regression coefficients are divided by the corresponding scale from \code{scales}, except intercept parameters. These are shifted by corresponding linear combination of \code{centers} given by rescaled other coefficients. Parameters for numeric outcomes also have to account for the rescaling of the outcome variable. For ordinal outcomes, intercepts are not included among the regressors, this is carried by the ordered intercepts \code{c_ord} and its reparametrizations. Also random effects including the covariance matrix (and related parameters) are rescaled. More details in the paper.
}
\value{
Returns an object equivalent to \code{draws}.
}
\references{
Vávra, J., Komárek, A., Grün, B., Malsiner-Walli, G. Clusterwise multivariate regression of mixed-type panel data. Stat Comput 34, 46 (2024). https://doi.org/10.1007/s11222-023-10304-5
}
\author{
Jan Vávra
}
\seealso{
\code{\link{clustGLMM}}
}
\examples{

### Get some longitudinal mixed type data
### here generated by generate_longitudinal_mixed_type_data()
data("longitudinal_mixed_type_data")

# Examples here demonstrate the basic use of the available functions on a toy dataset.
# For more thorough analyses see the vignettes.


###-----------------------------------------------------------------------------
###     Prepare inputs for clustGLMM()
###-----------------------------------------------------------------------------
id <- "i"

family <- c("num", "poi", "bin", "ord", "cat")
names(family) <- Ys <- c("ynum", "ypoi", "ybin", "yord", "ycat")

## Create list of formulae for each outcome
formula <- list()
for(y in Ys){
  formula[[y]] <- list()
  formula[[y]]$fixed <- ~ f
  formula[[y]]$group <- ~ x
  formula[[y]]$random <- ~ 1
  formula[[y]]$offset <- ""
}
formula
# intercept term will be group-specific 
# since it appears in $group which has a preference over $fixed

## All other inputs need to be prepared,
## unless the default values are desired (varying, save, howsave, param, tuning).
## See the vignette for how these can be adjusted.

###-----------------------------------------------------------------------------
###     Calling clustGLMM()
###-----------------------------------------------------------------------------

Gmax <- 5     # maximal number of mixture components to be considered
nchains <- 2  # number of chains to be sampled 

## No inits specified
set.seed(123456789)
mcmc <- clustGLMM(formula = formula, id = id, family = family, 
                  data = longitudinal_mixed_type_data, 
                  G = Gmax, iter = 100, nchains = nchains,
                  standardize = TRUE)
# rescale_list or rescale_matrix have already been called inside clustGLMM!!!
                  
###-----------------------------------------------------------------------------
###     !!! An illustrative use, does not make sense!!! 
###          !!! output is already scaled back !!!
###-----------------------------------------------------------------------------
some_draws <- vector("list", nchains)
for (ch in seq_len(nchains)) {
  some_draws[[ch]] <-
      rescale_matrix(mcmc$draws[[ch]], mcmc$settings[[ch]],
                     mcmc$family, mcmc$lfixnames, mcmc$lgrpnames, mcmc$lrannames,
                     mcmc$Kord, mcmc$Kcat, mcmc$centers, mcmc$scales)
}
}
\keyword{internal}
